backend/requirements.txt
# FastAPI and Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation & Environment
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# CORS & Middleware
python-multipart==0.0.6

# Development
pytest==7.4.4
httpx==0.26.0

# Email & Security Services
sendgrid==6.11.0
